requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
torch==2.1.0
faiss-cpu==1.7.4
numpy==1.24.3
